+++ packages/adminui-pub/CHANGELOG.md
@@ -9,10 +9,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - **CRYTICAL** EntityGridPanel: prevent memory leak by destroying grid popup menu
 - **CRYTICAL** Multifilter: prevent memory leak by destroying all Multifilter panels on destroy
 - **CRYTICAL** Multifilter: prevent memory leak for filter label toolbar (moved to data-qtip)
-- optimization: prevent creation of div's for floating elements shadow - all shadows are made using pure CSS 
 
 ### Changed
-- `Ext.picker.Date` now use CSS shadow instead of creation div element 
+- `Ext.picker.Date` now use CSS shadow instead of creation div element
+- optimization: prevent creation of div's for floating elements shadow - all shadows are made using pure CSS
+- optimization: lazy loading of components required by form / shortcut editor (`app/view/CommandBuilder/*`)
+- optimization: remove unused `AdvancedTextArea` control  
 
 ## [5.5.0]
 ### Fixed